SPTCH-3413: fix odd total timer reset in file_do()
The total timer is properly reset in MSTATE_INIT.
MSTATE_CONNECT starts with resetting the timer that is a start point for
further multi states.
If file://, MSTATE_DO calls file_do() that should not reset the total timer.
Otherwise, the total time is always less than the pre-transfer and the start
transfer times.